requirements.txt
numpy
pillow
paddlepaddle==2.5.2
lac
requests
timm
six